chore(ACL): move splitSubjects to utils [YTFRONT-5158]
